+++ src/main/php/lang/XPClass.class.php
@@ -134,6 +134,8 @@ class XPClass extends Type {
   public function newInstance($value= null) {
     if ($this->_reflect->isInterface()) {
       throw new IllegalAccessException('Cannot instantiate interfaces ('.$this->name.')');
+    } else if ($this->_reflect->isTrait()) {
+      throw new IllegalAccessException('Cannot instantiate traits ('.$this->name.')');
     } else if ($this->_reflect->isAbstract()) {
       throw new IllegalAccessException('Cannot instantiate abstract classes ('.$this->name.')');
     }
@@ -402,7 +404,7 @@ class XPClass extends Type {
   public function isInstance($obj) {
     return is($this->name, $obj);
   }
-  
+
   /**
    * Determines if this XPClass object represents an interface type.
    *
@@ -413,6 +415,15 @@ class XPClass extends Type {
   }
 
   /**
+   * Determines if this XPClass object represents a trait type.
+   *
+   * @return  bool
+   */
+  public function isTrait() {
+    return $this->_reflect->isTrait();
+  }
+
+  /**
    * Determines if this XPClass object represents an interface type.
    *
    * @return  bool
@@ -420,7 +431,22 @@ class XPClass extends Type {
   public function isEnum() {
     return class_exists('lang\Enum', false) && $this->_reflect->isSubclassOf('lang\Enum');
   }
-  
+
+  /**
+   * Retrieve traits this class uses
+   *
+   * @return  lang.XPClass[]
+   */
+  public function getTraits() {
+    $r= [];
+    foreach ($this->_reflect->getTraits() as $used) {
+      if (0 !== strncmp($used->getName(), '__', 2)) {
+        $r[]= new self($used);
+      }
+    }
+    return $r;
+  }
+
   /**
    * Retrieve interfaces this class implements
    *
@@ -429,7 +455,7 @@ class XPClass extends Type {
   public function getInterfaces() {
     $r= [];
     foreach ($this->_reflect->getInterfaces() as $iface) {
-      $r[]= new self($iface->getName());
+      $r[]= new self($iface);
     }
     return $r;
   }